#114 Amelioration IA + review architecture projet + fixes
