+++ Assets/Resources/Scripts/Utility.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e552abfb6e1c60544a119e76377de7ea
+folderAsset: yes
+timeCreated: 1458205793
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
